Feature/optional wait contracts

+++ scripts/keeper.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-rm -f /keeper-contracts/artifacts/ready
+[ "${LOCAL_CONTRACTS}" = "true" ] && rm -f /keeper-contracts/artifacts/ready
 ganache-cli -d -b ${BLOCK_TIME} --hostname "${LISTEN_ADDRESS}" --port "${LISTEN_PORT}" &
 
 sleep 2
@@ -8,6 +8,7 @@ sleep 2
 truffle migrate
 
 # Flag to indicate contracts are ready
-touch /keeper-contracts/artifacts/ready
+[ "${LOCAL_CONTRACTS}" = "true" ] && touch /keeper-contracts/artifacts/ready
 
 tail -f /dev/null
+
